+++ example/example.client.ts
@@ -11,31 +11,15 @@ type Type1 = {
 
 /** get /v1/user/retrieve */
 type GetV1UserRetrievePositiveVariant1 = {
-  status: "success";
-  data: {
-    id: number;
-    name: string;
-    features: Type1[];
-  };
+  id: number;
+  name: string;
+  features: Type1[];
 };
-
-/** get /v1/user/retrieve */
-interface GetV1UserRetrievePositiveResponseVariants {
-  200: GetV1UserRetrievePositiveVariant1;
-}
 
 /** get /v1/user/retrieve */
 type GetV1UserRetrieveNegativeVariant1 = {
-  status: "error";
-  error: {
-    message: string;
-  };
+  message: string;
 };
-
-/** get /v1/user/retrieve */
-interface GetV1UserRetrieveNegativeResponseVariants {
-  400: GetV1UserRetrieveNegativeVariant1;
-}
 
 /** head /v1/user/retrieve */
 type HeadV1UserRetrieveInput = {
@@ -47,17 +31,7 @@ type HeadV1UserRetrieveInput = {
 type HeadV1UserRetrievePositiveVariant1 = undefined;
 
 /** head /v1/user/retrieve */
-interface HeadV1UserRetrievePositiveResponseVariants {
-  200: HeadV1UserRetrievePositiveVariant1;
-}
-
-/** head /v1/user/retrieve */
 type HeadV1UserRetrieveNegativeVariant1 = undefined;
-
-/** head /v1/user/retrieve */
-interface HeadV1UserRetrieveNegativeResponseVariants {
-  400: HeadV1UserRetrieveNegativeVariant1;
-}
 
 /** delete /v1/user/:id/remove */
 type DeleteV1UserIdRemoveInput = {
@@ -69,17 +43,7 @@ type DeleteV1UserIdRemoveInput = {
 type DeleteV1UserIdRemovePositiveVariant1 = undefined;
 
 /** delete /v1/user/:id/remove */
-interface DeleteV1UserIdRemovePositiveResponseVariants {
-  204: DeleteV1UserIdRemovePositiveVariant1;
-}
-
-/** delete /v1/user/:id/remove */
 type DeleteV1UserIdRemoveNegativeVariant1 = undefined;
-
-/** delete /v1/user/:id/remove */
-interface DeleteV1UserIdRemoveNegativeResponseVariants {
-  404: DeleteV1UserIdRemoveNegativeVariant1;
-}
 
 /** patch /v1/user/:id */
 type PatchV1UserIdInput = {
@@ -93,34 +57,15 @@ type PatchV1UserIdInput = {
 
 /** patch /v1/user/:id */
 type PatchV1UserIdPositiveVariant1 = {
-  status: "success";
-  data: {
-    name: string;
-    /** account creation date */
-    createdAt: string;
-  };
+  name: string;
+  /** account creation date */
+  createdAt: string;
 };
-
-/** patch /v1/user/:id */
-interface PatchV1UserIdPositiveResponseVariants {
-  200: PatchV1UserIdPositiveVariant1;
-}
 
 /** patch /v1/user/:id */
 type PatchV1UserIdNegativeVariant1 = {
-  status: "error";
-  error: {
-    message: string;
-  };
+  message: string;
 };
-
-/** patch /v1/user/:id */
-interface PatchV1UserIdNegativeResponseVariants {
-  429: PatchV1UserIdNegativeVariant1;
-  401: PatchV1UserIdNegativeVariant1;
-  400: PatchV1UserIdNegativeVariant1;
-  404: PatchV1UserIdNegativeVariant1;
-}
 
 /** post /v1/user/create */
 type PostV1UserCreateInput = {
@@ -137,11 +82,6 @@ type PostV1UserCreatePositiveVariant1 = {
 };
 
 /** post /v1/user/create */
-interface PostV1UserCreatePositiveResponseVariants {
-  201: PostV1UserCreatePositiveVariant1;
-}
-
-/** post /v1/user/create */
 type PostV1UserCreateNegativeVariant1 = {
   status: "exists";
   id: number;
@@ -152,13 +92,6 @@ type PostV1UserCreateNegativeVariant2 = {
   status: "error";
   reason: string;
 };
-
-/** post /v1/user/create */
-interface PostV1UserCreateNegativeResponseVariants {
-  409: PostV1UserCreateNegativeVariant1;
-  400: PostV1UserCreateNegativeVariant2;
-  500: PostV1UserCreateNegativeVariant2;
-}
 
 /** query /v1/user/list */
 type QueryV1UserListInput = {
@@ -172,17 +105,7 @@ type QueryV1UserListPositiveVariant1 = {
 }[];
 
 /** query /v1/user/list */
-interface QueryV1UserListPositiveResponseVariants {
-  200: QueryV1UserListPositiveVariant1;
-}
-
-/** query /v1/user/list */
 type QueryV1UserListNegativeVariant1 = string;
-
-/** query /v1/user/list */
-interface QueryV1UserListNegativeResponseVariants {
-  400: QueryV1UserListNegativeVariant1;
-}
 
 /** post /v1/login */
 type PostV1LoginInput = {
@@ -192,29 +115,13 @@ type PostV1LoginInput = {
 
 /** post /v1/login */
 type PostV1LoginPositiveVariant1 = {
-  status: "success";
-  data: {
-    message: string;
-  };
+  message: string;
 };
-
-/** post /v1/login */
-interface PostV1LoginPositiveResponseVariants {
-  200: PostV1LoginPositiveVariant1;
-}
 
 /** post /v1/login */
 type PostV1LoginNegativeVariant1 = {
-  status: "error";
-  error: {
-    message: string;
-  };
+  message: string;
 };
-
-/** post /v1/login */
-interface PostV1LoginNegativeResponseVariants {
-  400: PostV1LoginNegativeVariant1;
-}
 
 /** get /v1/avatar/send */
 type GetV1AvatarSendInput = {
@@ -225,17 +132,7 @@ type GetV1AvatarSendInput = {
 type GetV1AvatarSendPositiveVariant1 = string;
 
 /** get /v1/avatar/send */
-interface GetV1AvatarSendPositiveResponseVariants {
-  200: GetV1AvatarSendPositiveVariant1;
-}
-
-/** get /v1/avatar/send */
 type GetV1AvatarSendNegativeVariant1 = string;
-
-/** get /v1/avatar/send */
-interface GetV1AvatarSendNegativeResponseVariants {
-  400: GetV1AvatarSendNegativeVariant1;
-}
 
 /** head /v1/avatar/send */
 type HeadV1AvatarSendInput = {
@@ -246,17 +143,7 @@ type HeadV1AvatarSendInput = {
 type HeadV1AvatarSendPositiveVariant1 = undefined;
 
 /** head /v1/avatar/send */
-interface HeadV1AvatarSendPositiveResponseVariants {
-  200: HeadV1AvatarSendPositiveVariant1;
-}
-
-/** head /v1/avatar/send */
 type HeadV1AvatarSendNegativeVariant1 = undefined;
-
-/** head /v1/avatar/send */
-interface HeadV1AvatarSendNegativeResponseVariants {
-  400: HeadV1AvatarSendNegativeVariant1;
-}
 
 /** get /v1/avatar/stream */
 type GetV1AvatarStreamInput = {
@@ -267,17 +154,7 @@ type GetV1AvatarStreamInput = {
 type GetV1AvatarStreamPositiveVariant1 = Blob;
 
 /** get /v1/avatar/stream */
-interface GetV1AvatarStreamPositiveResponseVariants {
-  200: GetV1AvatarStreamPositiveVariant1;
-}
-
-/** get /v1/avatar/stream */
 type GetV1AvatarStreamNegativeVariant1 = string;
-
-/** get /v1/avatar/stream */
-interface GetV1AvatarStreamNegativeResponseVariants {
-  400: GetV1AvatarStreamNegativeVariant1;
-}
 
 /** head /v1/avatar/stream */
 type HeadV1AvatarStreamInput = {
@@ -288,17 +165,7 @@ type HeadV1AvatarStreamInput = {
 type HeadV1AvatarStreamPositiveVariant1 = undefined;
 
 /** head /v1/avatar/stream */
-interface HeadV1AvatarStreamPositiveResponseVariants {
-  200: HeadV1AvatarStreamPositiveVariant1;
-}
-
-/** head /v1/avatar/stream */
 type HeadV1AvatarStreamNegativeVariant1 = undefined;
-
-/** head /v1/avatar/stream */
-interface HeadV1AvatarStreamNegativeResponseVariants {
-  400: HeadV1AvatarStreamNegativeVariant1;
-}
 
 /** post /v1/avatar/upload */
 type PostV1AvatarUploadInput = Omit<
@@ -314,62 +181,30 @@ type PostV1AvatarUploadInput = Omit<
 
 /** post /v1/avatar/upload */
 type PostV1AvatarUploadPositiveVariant1 = {
-  status: "success";
-  data: {
-    name: string;
-    size: number;
-    mime: string;
-    hash: string;
-    otherInputs: Record<string, any>;
-  };
+  name: string;
+  size: number;
+  mime: string;
+  hash: string;
+  otherInputs: Record<string, any>;
 };
-
-/** post /v1/avatar/upload */
-interface PostV1AvatarUploadPositiveResponseVariants {
-  200: PostV1AvatarUploadPositiveVariant1;
-}
 
 /** post /v1/avatar/upload */
 type PostV1AvatarUploadNegativeVariant1 = {
-  status: "error";
-  error: {
-    message: string;
-  };
+  message: string;
 };
-
-/** post /v1/avatar/upload */
-interface PostV1AvatarUploadNegativeResponseVariants {
-  400: PostV1AvatarUploadNegativeVariant1;
-}
 
 /** post /v1/avatar/raw */
 type PostV1AvatarRawInput = Blob;
 
 /** post /v1/avatar/raw */
 type PostV1AvatarRawPositiveVariant1 = {
-  status: "success";
-  data: {
-    length: number;
-  };
+  length: number;
 };
-
-/** post /v1/avatar/raw */
-interface PostV1AvatarRawPositiveResponseVariants {
-  200: PostV1AvatarRawPositiveVariant1;
-}
 
 /** post /v1/avatar/raw */
 type PostV1AvatarRawNegativeVariant1 = {
-  status: "error";
-  error: {
-    message: string;
-  };
+  message: string;
 };
-
-/** post /v1/avatar/raw */
-interface PostV1AvatarRawNegativeResponseVariants {
-  400: PostV1AvatarRawNegativeVariant1;
-}
 
 /** get /v1/events/stream */
 type GetV1EventsStreamInput = {
@@ -386,17 +221,7 @@ type GetV1EventsStreamPositiveVariant1 = {
 };
 
 /** get /v1/events/stream */
-interface GetV1EventsStreamPositiveResponseVariants {
-  200: GetV1EventsStreamPositiveVariant1;
-}
-
-/** get /v1/events/stream */
 type GetV1EventsStreamNegativeVariant1 = string;
-
-/** get /v1/events/stream */
-interface GetV1EventsStreamNegativeResponseVariants {
-  400: GetV1EventsStreamNegativeVariant1;
-}
 
 /** head /v1/events/stream */
 type HeadV1EventsStreamInput = {
@@ -408,17 +233,7 @@ type HeadV1EventsStreamInput = {
 type HeadV1EventsStreamPositiveVariant1 = undefined;
 
 /** head /v1/events/stream */
-interface HeadV1EventsStreamPositiveResponseVariants {
-  200: HeadV1EventsStreamPositiveVariant1;
-}
-
-/** head /v1/events/stream */
 type HeadV1EventsStreamNegativeVariant1 = undefined;
-
-/** head /v1/events/stream */
-interface HeadV1EventsStreamNegativeResponseVariants {
-  400: HeadV1EventsStreamNegativeVariant1;
-}
 
 /** post /v1/forms/feedback */
 type PostV1FormsFeedbackInput = {
@@ -429,29 +244,13 @@ type PostV1FormsFeedbackInput = {
 
 /** post /v1/forms/feedback */
 type PostV1FormsFeedbackPositiveVariant1 = {
-  status: "success";
-  data: {
-    crc: number;
-  };
+  crc: number;
 };
-
-/** post /v1/forms/feedback */
-interface PostV1FormsFeedbackPositiveResponseVariants {
-  200: PostV1FormsFeedbackPositiveVariant1;
-}
 
 /** post /v1/forms/feedback */
 type PostV1FormsFeedbackNegativeVariant1 = {
-  status: "error";
-  error: {
-    message: string;
-  };
+  message: string;
 };
-
-/** post /v1/forms/feedback */
-interface PostV1FormsFeedbackNegativeResponseVariants {
-  400: PostV1FormsFeedbackNegativeVariant1;
-}
 
 /** get /v2/users/list */
 type GetV2UsersListInput = {
@@ -465,39 +264,23 @@ type GetV2UsersListInput = {
 
 /** get /v2/users/list */
 type GetV2UsersListPositiveVariant1 = {
-  status: "success";
-  data: {
-    /** Page of users */
-    users: {
-      name: string;
-      role: "manager" | "operator" | "admin";
-    }[];
-    /** Total number of users */
-    total: number;
-    /** Page size used */
-    limit: number;
-    /** Offset used */
-    offset: number;
-  };
+  /** Page of users */
+  users: {
+    name: string;
+    role: "manager" | "operator" | "admin";
+  }[];
+  /** Total number of users */
+  total: number;
+  /** Page size used */
+  limit: number;
+  /** Offset used */
+  offset: number;
 };
-
-/** get /v2/users/list */
-interface GetV2UsersListPositiveResponseVariants {
-  200: GetV2UsersListPositiveVariant1;
-}
 
 /** get /v2/users/list */
 type GetV2UsersListNegativeVariant1 = {
-  status: "error";
-  error: {
-    message: string;
-  };
+  message: string;
 };
-
-/** get /v2/users/list */
-interface GetV2UsersListNegativeResponseVariants {
-  400: GetV2UsersListNegativeVariant1;
-}
 
 /** head /v2/users/list */
 type HeadV2UsersListInput = {
@@ -513,19 +296,7 @@ type HeadV2UsersListInput = {
 type HeadV2UsersListPositiveVariant1 = undefined;
 
 /** head /v2/users/list */
-interface HeadV2UsersListPositiveResponseVariants {
-  200: HeadV2UsersListPositiveVariant1;
-}
-
-/** head /v2/users/list */
 type HeadV2UsersListNegativeVariant1 = undefined;
-
-/** head /v2/users/list */
-interface HeadV2UsersListNegativeResponseVariants {
-  400: HeadV2UsersListNegativeVariant1;
-}
-
-type SomeOf<T> = T[keyof T];
 
 export type Path =
   | "/v1/user/retrieve"
@@ -568,146 +339,136 @@ export interface Input {
 }
 
 export interface PositiveResponse {
-  "get /v1/user/retrieve": SomeOf<GetV1UserRetrievePositiveResponseVariants>;
-  "head /v1/user/retrieve": SomeOf<HeadV1UserRetrievePositiveResponseVariants>;
-  "delete /v1/user/:id/remove": SomeOf<DeleteV1UserIdRemovePositiveResponseVariants>;
-  "patch /v1/user/:id": SomeOf<PatchV1UserIdPositiveResponseVariants>;
-  "post /v1/user/create": SomeOf<PostV1UserCreatePositiveResponseVariants>;
-  "query /v1/user/list": SomeOf<QueryV1UserListPositiveResponseVariants>;
-  "post /v1/login": SomeOf<PostV1LoginPositiveResponseVariants>;
+  "get /v1/user/retrieve": GetV1UserRetrievePositiveVariant1;
+  "head /v1/user/retrieve": HeadV1UserRetrievePositiveVariant1;
+  "delete /v1/user/:id/remove": DeleteV1UserIdRemovePositiveVariant1;
+  "patch /v1/user/:id": PatchV1UserIdPositiveVariant1;
+  "post /v1/user/create": PostV1UserCreatePositiveVariant1;
+  "query /v1/user/list": QueryV1UserListPositiveVariant1;
+  "post /v1/login": PostV1LoginPositiveVariant1;
   /** @deprecated */
-  "get /v1/avatar/send": SomeOf<GetV1AvatarSendPositiveResponseVariants>;
+  "get /v1/avatar/send": GetV1AvatarSendPositiveVariant1;
   /** @deprecated */
-  "head /v1/avatar/send": SomeOf<HeadV1AvatarSendPositiveResponseVariants>;
-  "get /v1/avatar/stream": SomeOf<GetV1AvatarStreamPositiveResponseVariants>;
-  "head /v1/avatar/stream": SomeOf<HeadV1AvatarStreamPositiveResponseVariants>;
-  "post /v1/avatar/upload": SomeOf<PostV1AvatarUploadPositiveResponseVariants>;
-  "post /v1/avatar/raw": SomeOf<PostV1AvatarRawPositiveResponseVariants>;
-  "get /v1/events/stream": SomeOf<GetV1EventsStreamPositiveResponseVariants>;
-  "head /v1/events/stream": SomeOf<HeadV1EventsStreamPositiveResponseVariants>;
-  "post /v1/forms/feedback": SomeOf<PostV1FormsFeedbackPositiveResponseVariants>;
-  "get /v2/users/list": SomeOf<GetV2UsersListPositiveResponseVariants>;
-  "head /v2/users/list": SomeOf<HeadV2UsersListPositiveResponseVariants>;
+  "head /v1/avatar/send": HeadV1AvatarSendPositiveVariant1;
+  "get /v1/avatar/stream": GetV1AvatarStreamPositiveVariant1;
+  "head /v1/avatar/stream": HeadV1AvatarStreamPositiveVariant1;
+  "post /v1/avatar/upload": PostV1AvatarUploadPositiveVariant1;
+  "post /v1/avatar/raw": PostV1AvatarRawPositiveVariant1;
+  "get /v1/events/stream": GetV1EventsStreamPositiveVariant1;
+  "head /v1/events/stream": HeadV1EventsStreamPositiveVariant1;
+  "post /v1/forms/feedback": PostV1FormsFeedbackPositiveVariant1;
+  "get /v2/users/list": GetV2UsersListPositiveVariant1;
+  "head /v2/users/list": HeadV2UsersListPositiveVariant1;
 }
 
 export interface NegativeResponse {
-  "get /v1/user/retrieve": SomeOf<GetV1UserRetrieveNegativeResponseVariants>;
-  "head /v1/user/retrieve": SomeOf<HeadV1UserRetrieveNegativeResponseVariants>;
-  "delete /v1/user/:id/remove": SomeOf<DeleteV1UserIdRemoveNegativeResponseVariants>;
-  "patch /v1/user/:id": SomeOf<PatchV1UserIdNegativeResponseVariants>;
-  "post /v1/user/create": SomeOf<PostV1UserCreateNegativeResponseVariants>;
-  "query /v1/user/list": SomeOf<QueryV1UserListNegativeResponseVariants>;
-  "post /v1/login": SomeOf<PostV1LoginNegativeResponseVariants>;
+  "get /v1/user/retrieve": GetV1UserRetrieveNegativeVariant1;
+  "head /v1/user/retrieve": HeadV1UserRetrieveNegativeVariant1;
+  "delete /v1/user/:id/remove": DeleteV1UserIdRemoveNegativeVariant1;
+  "patch /v1/user/:id": PatchV1UserIdNegativeVariant1;
+  "post /v1/user/create": PostV1UserCreateNegativeVariant1 | PostV1UserCreateNegativeVariant2;
+  "query /v1/user/list": QueryV1UserListNegativeVariant1;
+  "post /v1/login": PostV1LoginNegativeVariant1;
   /** @deprecated */
-  "get /v1/avatar/send": SomeOf<GetV1AvatarSendNegativeResponseVariants>;
+  "get /v1/avatar/send": GetV1AvatarSendNegativeVariant1;
   /** @deprecated */
-  "head /v1/avatar/send": SomeOf<HeadV1AvatarSendNegativeResponseVariants>;
-  "get /v1/avatar/stream": SomeOf<GetV1AvatarStreamNegativeResponseVariants>;
-  "head /v1/avatar/stream": SomeOf<HeadV1AvatarStreamNegativeResponseVariants>;
-  "post /v1/avatar/upload": SomeOf<PostV1AvatarUploadNegativeResponseVariants>;
-  "post /v1/avatar/raw": SomeOf<PostV1AvatarRawNegativeResponseVariants>;
-  "get /v1/events/stream": SomeOf<GetV1EventsStreamNegativeResponseVariants>;
-  "head /v1/events/stream": SomeOf<HeadV1EventsStreamNegativeResponseVariants>;
-  "post /v1/forms/feedback": SomeOf<PostV1FormsFeedbackNegativeResponseVariants>;
-  "get /v2/users/list": SomeOf<GetV2UsersListNegativeResponseVariants>;
-  "head /v2/users/list": SomeOf<HeadV2UsersListNegativeResponseVariants>;
+  "head /v1/avatar/send": HeadV1AvatarSendNegativeVariant1;
+  "get /v1/avatar/stream": GetV1AvatarStreamNegativeVariant1;
+  "head /v1/avatar/stream": HeadV1AvatarStreamNegativeVariant1;
+  "post /v1/avatar/upload": PostV1AvatarUploadNegativeVariant1;
+  "post /v1/avatar/raw": PostV1AvatarRawNegativeVariant1;
+  "get /v1/events/stream": GetV1EventsStreamNegativeVariant1;
+  "head /v1/events/stream": HeadV1EventsStreamNegativeVariant1;
+  "post /v1/forms/feedback": PostV1FormsFeedbackNegativeVariant1;
+  "get /v2/users/list": GetV2UsersListNegativeVariant1;
+  "head /v2/users/list": HeadV2UsersListNegativeVariant1;
 }
 
 export interface EncodedResponse {
-  "get /v1/user/retrieve": GetV1UserRetrievePositiveResponseVariants &
-    GetV1UserRetrieveNegativeResponseVariants;
-  "head /v1/user/retrieve": HeadV1UserRetrievePositiveResponseVariants &
-    HeadV1UserRetrieveNegativeResponseVariants;
-  "delete /v1/user/:id/remove": DeleteV1UserIdRemovePositiveResponseVariants &
-    DeleteV1UserIdRemoveNegativeResponseVariants;
-  "patch /v1/user/:id": PatchV1UserIdPositiveResponseVariants &
-    PatchV1UserIdNegativeResponseVariants;
-  "post /v1/user/create": PostV1UserCreatePositiveResponseVariants &
-    PostV1UserCreateNegativeResponseVariants;
-  "query /v1/user/list": QueryV1UserListPositiveResponseVariants &
-    QueryV1UserListNegativeResponseVariants;
-  "post /v1/login": PostV1LoginPositiveResponseVariants & PostV1LoginNegativeResponseVariants;
+  "get /v1/user/retrieve":
+    | { status: 200; discriminator: "success"; data: GetV1UserRetrievePositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: GetV1UserRetrieveNegativeVariant1 };
+  "head /v1/user/retrieve":
+    | { status: 200; discriminator: "success"; data: HeadV1UserRetrievePositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: HeadV1UserRetrieveNegativeVariant1 };
+  "delete /v1/user/:id/remove":
+    | { status: 204; discriminator: "success"; data: DeleteV1UserIdRemovePositiveVariant1 }
+    | { status: 404; discriminator: "error"; data: DeleteV1UserIdRemoveNegativeVariant1 };
+  "patch /v1/user/:id":
+    | { status: 200; discriminator: "success"; data: PatchV1UserIdPositiveVariant1 }
+    | {
+        status: 429 | 401 | 400 | 404;
+        discriminator: "error";
+        data: PatchV1UserIdNegativeVariant1;
+      };
+  "post /v1/user/create":
+    | { status: 201; discriminator: "success"; data: PostV1UserCreatePositiveVariant1 }
+    | { status: 409; discriminator: "error"; data: PostV1UserCreateNegativeVariant1 }
+    | { status: 400 | 500; discriminator: "error"; data: PostV1UserCreateNegativeVariant2 };
+  "query /v1/user/list":
+    | { status: 200; discriminator: "success"; data: QueryV1UserListPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: QueryV1UserListNegativeVariant1 };
+  "post /v1/login":
+    | { status: 200; discriminator: "success"; data: PostV1LoginPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: PostV1LoginNegativeVariant1 };
   /** @deprecated */
-  "get /v1/avatar/send": GetV1AvatarSendPositiveResponseVariants &
-    GetV1AvatarSendNegativeResponseVariants;
+  "get /v1/avatar/send":
+    | { status: 200; discriminator: "success"; data: GetV1AvatarSendPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: GetV1AvatarSendNegativeVariant1 };
   /** @deprecated */
-  "head /v1/avatar/send": HeadV1AvatarSendPositiveResponseVariants &
-    HeadV1AvatarSendNegativeResponseVariants;
-  "get /v1/avatar/stream": GetV1AvatarStreamPositiveResponseVariants &
-    GetV1AvatarStreamNegativeResponseVariants;
-  "head /v1/avatar/stream": HeadV1AvatarStreamPositiveResponseVariants &
-    HeadV1AvatarStreamNegativeResponseVariants;
-  "post /v1/avatar/upload": PostV1AvatarUploadPositiveResponseVariants &
-    PostV1AvatarUploadNegativeResponseVariants;
-  "post /v1/avatar/raw": PostV1AvatarRawPositiveResponseVariants &
-    PostV1AvatarRawNegativeResponseVariants;
-  "get /v1/events/stream": GetV1EventsStreamPositiveResponseVariants &
-    GetV1EventsStreamNegativeResponseVariants;
-  "head /v1/events/stream": HeadV1EventsStreamPositiveResponseVariants &
-    HeadV1EventsStreamNegativeResponseVariants;
-  "post /v1/forms/feedback": PostV1FormsFeedbackPositiveResponseVariants &
-    PostV1FormsFeedbackNegativeResponseVariants;
-  "get /v2/users/list": GetV2UsersListPositiveResponseVariants &
-    GetV2UsersListNegativeResponseVariants;
-  "head /v2/users/list": HeadV2UsersListPositiveResponseVariants &
-    HeadV2UsersListNegativeResponseVariants;
+  "head /v1/avatar/send":
+    | { status: 200; discriminator: "success"; data: HeadV1AvatarSendPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: HeadV1AvatarSendNegativeVariant1 };
+  "get /v1/avatar/stream":
+    | { status: 200; discriminator: "success"; data: GetV1AvatarStreamPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: GetV1AvatarStreamNegativeVariant1 };
+  "head /v1/avatar/stream":
+    | { status: 200; discriminator: "success"; data: HeadV1AvatarStreamPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: HeadV1AvatarStreamNegativeVariant1 };
+  "post /v1/avatar/upload":
+    | { status: 200; discriminator: "success"; data: PostV1AvatarUploadPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: PostV1AvatarUploadNegativeVariant1 };
+  "post /v1/avatar/raw":
+    | { status: 200; discriminator: "success"; data: PostV1AvatarRawPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: PostV1AvatarRawNegativeVariant1 };
+  "get /v1/events/stream":
+    | { status: 200; discriminator: "success"; data: GetV1EventsStreamPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: GetV1EventsStreamNegativeVariant1 };
+  "head /v1/events/stream":
+    | { status: 200; discriminator: "success"; data: HeadV1EventsStreamPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: HeadV1EventsStreamNegativeVariant1 };
+  "post /v1/forms/feedback":
+    | { status: 200; discriminator: "success"; data: PostV1FormsFeedbackPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: PostV1FormsFeedbackNegativeVariant1 };
+  "get /v2/users/list":
+    | { status: 200; discriminator: "success"; data: GetV2UsersListPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: GetV2UsersListNegativeVariant1 };
+  "head /v2/users/list":
+    | { status: 200; discriminator: "success"; data: HeadV2UsersListPositiveVariant1 }
+    | { status: 400; discriminator: "error"; data: HeadV2UsersListNegativeVariant1 };
 }
 
 export interface Response {
-  "get /v1/user/retrieve":
-    | PositiveResponse["get /v1/user/retrieve"]
-    | NegativeResponse["get /v1/user/retrieve"];
-  "head /v1/user/retrieve":
-    | PositiveResponse["head /v1/user/retrieve"]
-    | NegativeResponse["head /v1/user/retrieve"];
-  "delete /v1/user/:id/remove":
-    | PositiveResponse["delete /v1/user/:id/remove"]
-    | NegativeResponse["delete /v1/user/:id/remove"];
-  "patch /v1/user/:id":
-    | PositiveResponse["patch /v1/user/:id"]
-    | NegativeResponse["patch /v1/user/:id"];
-  "post /v1/user/create":
-    | PositiveResponse["post /v1/user/create"]
-    | NegativeResponse["post /v1/user/create"];
-  "query /v1/user/list":
-    | PositiveResponse["query /v1/user/list"]
-    | NegativeResponse["query /v1/user/list"];
-  "post /v1/login": PositiveResponse["post /v1/login"] | NegativeResponse["post /v1/login"];
+  "get /v1/user/retrieve": EncodedResponse["get /v1/user/retrieve"]["data"];
+  "head /v1/user/retrieve": EncodedResponse["head /v1/user/retrieve"]["data"];
+  "delete /v1/user/:id/remove": EncodedResponse["delete /v1/user/:id/remove"]["data"];
+  "patch /v1/user/:id": EncodedResponse["patch /v1/user/:id"]["data"];
+  "post /v1/user/create": EncodedResponse["post /v1/user/create"]["data"];
+  "query /v1/user/list": EncodedResponse["query /v1/user/list"]["data"];
+  "post /v1/login": EncodedResponse["post /v1/login"]["data"];
   /** @deprecated */
-  "get /v1/avatar/send":
-    | PositiveResponse["get /v1/avatar/send"]
-    | NegativeResponse["get /v1/avatar/send"];
+  "get /v1/avatar/send": EncodedResponse["get /v1/avatar/send"]["data"];
   /** @deprecated */
-  "head /v1/avatar/send":
-    | PositiveResponse["head /v1/avatar/send"]
-    | NegativeResponse["head /v1/avatar/send"];
-  "get /v1/avatar/stream":
-    | PositiveResponse["get /v1/avatar/stream"]
-    | NegativeResponse["get /v1/avatar/stream"];
-  "head /v1/avatar/stream":
-    | PositiveResponse["head /v1/avatar/stream"]
-    | NegativeResponse["head /v1/avatar/stream"];
-  "post /v1/avatar/upload":
-    | PositiveResponse["post /v1/avatar/upload"]
-    | NegativeResponse["post /v1/avatar/upload"];
-  "post /v1/avatar/raw":
-    | PositiveResponse["post /v1/avatar/raw"]
-    | NegativeResponse["post /v1/avatar/raw"];
-  "get /v1/events/stream":
-    | PositiveResponse["get /v1/events/stream"]
-    | NegativeResponse["get /v1/events/stream"];
-  "head /v1/events/stream":
-    | PositiveResponse["head /v1/events/stream"]
-    | NegativeResponse["head /v1/events/stream"];
-  "post /v1/forms/feedback":
-    | PositiveResponse["post /v1/forms/feedback"]
-    | NegativeResponse["post /v1/forms/feedback"];
-  "get /v2/users/list":
-    | PositiveResponse["get /v2/users/list"]
-    | NegativeResponse["get /v2/users/list"];
-  "head /v2/users/list":
-    | PositiveResponse["head /v2/users/list"]
-    | NegativeResponse["head /v2/users/list"];
+  "head /v1/avatar/send": EncodedResponse["head /v1/avatar/send"]["data"];
+  "get /v1/avatar/stream": EncodedResponse["get /v1/avatar/stream"]["data"];
+  "head /v1/avatar/stream": EncodedResponse["head /v1/avatar/stream"]["data"];
+  "post /v1/avatar/upload": EncodedResponse["post /v1/avatar/upload"]["data"];
+  "post /v1/avatar/raw": EncodedResponse["post /v1/avatar/raw"]["data"];
+  "get /v1/events/stream": EncodedResponse["get /v1/events/stream"]["data"];
+  "head /v1/events/stream": EncodedResponse["head /v1/events/stream"]["data"];
+  "post /v1/forms/feedback": EncodedResponse["post /v1/forms/feedback"]["data"];
+  "get /v2/users/list": EncodedResponse["get /v2/users/list"]["data"];
+  "head /v2/users/list": EncodedResponse["head /v2/users/list"]["data"];
 }
 
 export type Request = keyof Input;
@@ -752,7 +513,7 @@ export type Implementation<T extends Record<string, unknown>> = (
   path: string,
   params: Record<string, any>,
   ctx?: T,
-) => Promise<any>;
+) => Promise<{ status: number; data: any }>;
 
 type Pagination = { nextCursor: string | null } | { total: number; limit: number; offset: number };
 
@@ -788,24 +549,37 @@ const defaultImplementation: Implementation<DefaultContext> = async (method, pat
   };
   if (ctx?.override) init = ctx.override(init);
   const response = await fetch(new URL(`${path}${searchParams}`, "http://localhost:8090"), init);
+  const { status } = response;
   const contentType = response.headers.get("content-type");
-  if (!contentType) return;
-  if (contentType.startsWith("application/json")) return response.json();
-  if (contentType.startsWith("text/")) return response.text();
-  return response.blob();
+  if (!contentType) return { status, data: undefined };
+  const data = await (contentType.startsWith("application/json")
+    ? response.json()
+    : contentType.startsWith("text/")
+      ? response.text()
+      : response.blob());
+  return { status, data };
 };
 
 export class Client<T extends Record<string, unknown> = DefaultContext> {
   public constructor(
     protected readonly implementation: Implementation<T> = defaultImplementation,
   ) {}
-  public provide<K extends Request>(request: K, params: Input[K], ctx?: T): Promise<Response[K]> {
+  public async provide<K extends Request>(
+    request: K,
+    params: Input[K],
+    ctx?: T,
+  ): Promise<EncodedResponse[K]> {
     const [method, path] = parseRequest(request);
-    return this.implementation(method, ...substitute(path, params), ctx);
+    const { status, data } = await this.implementation(method, ...substitute(path, params), ctx);
+    const discriminator = Client.discriminate(status);
+    return { status, data, discriminator } as EncodedResponse[K];
   }
   public static hasMore(response: Pagination): boolean {
     if ("nextCursor" in response) return response.nextCursor !== null;
     return response.offset + response.limit < response.total;
+  }
+  public static discriminate(status: number): "success" | "error" {
+    return status < 400 ? "success" : "error";
   }
 }
 
@@ -833,6 +607,9 @@ export class Subscription<
 // Usage example:
 /*
 const client = new Client();
-client.provide("get /v1/user/retrieve", { id: "10" });
+const { status, discriminator, data } = await client.provide("get /v1/user/retrieve", { id: "10" });
+if (status === 200)
+  console.log(data.name); // success
+else if (status === 400 || discriminator === "error") console.error(data.message); // error
 new Subscription("get /v1/events/stream", {}).on("time", (time) => {});
 */